+++ src/main/java/calculator/View/OutputView.java
@@ -1,7 +1,8 @@
 package calculator.View;
 import static calculator.Constants.ViewConstants.END;
 public class OutputView {
-    public static void userOutput(int result){
+    public static void userOutput(long result){
         System.out.println(END.getMessage() + result);
     }
+
 }
